Move instrumentation to debug module
