+++ frontend/catalyst/debug/__init__.py
@@ -20,6 +20,7 @@ from catalyst.debug.compiler_functions import (
     get_cmain,
     print_compilation_stage,
 )
+from catalyst.debug.instruments import instrumentation
 from catalyst.debug.printing import (  # pylint: disable=redefined-builtin
     print,
     print_memref,
@@ -32,4 +33,5 @@ __all__ = (
     "print_compilation_stage",
     "get_cmain",
     "compile_from_mlir",
+    "instrumentation",
 )
