+++ frontend/test/lit/test_instrumentation_console.py
@@ -18,7 +18,8 @@
 import numpy as np
 import pennylane as qml
 
-from catalyst import instrumentation, qjit
+from catalyst import qjit
+from catalyst.debug import instrumentation
 
 num_layers = 1
 num_wires = 1
